Added a task to update the terminating-redirect url in the urls table.
Also renamed in inc/languages/english/admin the file
config_duplicate_link_warner.lang.php to duplicate_link_warner.lang.php
so it has the same name as the file in the parent directory, because in
the task file, which language file is loaded by a call to $lang->load()
depends on whether the task is called directly by an admin or as part of
a pseudo-cron system run at the end of ordinary user page accesses. This
means that the 'dlw_task_ran' message had to be present in both files.