Bugfix (update matching_urls_in_post) plus @todos for further fixes
Update matching_urls_in_post when matching_urls is updated. Steps to
reproduce (now fixed) bug:

1. Open an edit window for a new thread.
2. Add to the draft a URL present in a forum post known to contain
multiple URLs. Wait for the warning pop-up to appear.
3. Add to the draft another URL from that forum post.
4. Click on the "Show more" button in the warning pop-up.

Result: The pop-up will not expand, and the browser's Javascript console
will indicate an error which is essentially due to the
post.matching_urls_in_post array containing fewer entries than the
post.matching_urls array: when an entry is added to matching_urls based
on a match with all_urls, a corresponding entry is not added to
matching_urls_in_post.