Fixed a bunch of bugs/issues and made some minor additions.
The bugs/issues fixed, and minor additions made, include:

1. Adding the "t" parameter as one to ignore on YouTube.

2. Matching normalised URLs against not just the `url_term_norm` column
but also the `url_norm` column.

3. Returning only unique rows when matching urls in the DB.

4. Adding a missing TABLE_PREFIX to replace a hard-coded 'mybb_'.

5. Adding various constants.

6. HTML-decoding target URLs discovered via meta canonical tags and HTML
redirects.

7. Adding an entry to the array in dlw_get_url_term_redirs_auto().

8. Initialising the "use_head_method" parameter based on the relevant
constants.

9. Amending and adding to the defensive termination of the main loop in
dlw_get_url_term_redirs() such that it terminates based on:

   9.1. A roughly-determined (and very generous) maximum iteration
        count.

   9.2. A maximum run time of
        dlw_max_allowable_redirect_resolution_runtime_secs seconds.

   9.3. A  maximum per-url redirect count of
        dlw_max_allowable_redirects_for_a_url.

10. Reversed the order in which the $urls2 and $deferred_urls arrays are
merged in dlw_get_url_term_redirs() to ensure that as many redirects as
possible for each starting URL are followed in case the main loop is
terminated by timeout - i.e., that we do not waste time querying initial
redirects for URLs which we don't have time to follow to the end; we
"focus" on one (set of) URLs at a time when following redirects (to the
end).

11. Working around, in dlw_clean_up_dangling_urls(), a MySQL error
(which does not affect MariaDB).

12. Correcting in the SQL conditions in dlw_get_and_store_terms() a
reference to `last_term_try` into a reference to `term_tries`.

13. Reopening the DB connection in dlw_get_and_store_terms() in case it
"goes away", as it sometimes does on our testing host, Hostgator.

14. Misc other stuff including adding @todo and other comments.