Bugfixes: synced client-side JS with server-side changes, plus another.
1. The changes to server-side code, in particular the new feature
whereby links which resolve to the same terminating link are treated as
identical, required changes to be made to client-side code that had not
until this commit been made. The problem in particular that this had led
to was that entering a link into the editor which had matches, and then
replacing it with a superficially different but functionally identical
(due to same termination) link, and then changing it back again, could
lead the warning summary box to display a different (smaller) set of
posts the second time around than the first.

2. Fixed the XHR call so that existing post IDs and edit times are fully
transmitted to the server (which involved pluralising key 'pid' to
'pids' as the unpluralised version seemed to be conflicting with MyBB
core code).