Two changes: closed a detection hole, and stored best term redir
1. Closed the hole which allowed for some matches to slip through the
cracks: URLs that did not yet occur in any post but that redirected to
URLs which *did* occur in posts were not caught as matches for those
existing URLs. Closed this hole by querying for redirects in the AJAX
call and storing missing URLs into the DB before checking for matches.

2. When errors occur in resolving the terminating redirect for a URL,
store the furtherest redirected-to URL that we have and set got_term to
0 in the DB.

3. Always set term_tries to 1 and last_term_try to current time when
inserting URLs in to DB, regardless of what we set got_term to.