Added a status display in the plugin's panel in the admin plugin console.
The status displays the number of posts with unextracted links versus
total posts, and the number of links for which redirects have not been
resolved, both the ones which have not been attempted yet as well as
those which have been and in which the attempt failed, versus total
links. It provides the user with a means to run processes which do
what's necessary to extract/resolve links.