Decided that we don't need to check robots.txt
Why not? Because that file is primarily intended to stop search engines
from indexing certain content. We are not indexing content for searches
or for any other purpose: we are not indexing content at all. We are
simply trying to work out where a link terminates after all redirects of
all types are resolved.