Fix compatibility with other parser-calling plugins
Incompatibility as reported by Omar G., specifically w.r.t. his OUGC
Post Character Count Enhancement plugin, which calls the parser in its
`ougc_countchars()` function, which itself is called by a hook-in to the
`datahandler_post_validate_post` hook. Link Tools mistakenly took this
parser call to be that for the post.

The fix ensures that Link Tools is never mistaken about when to insert
preview placeholders into a post.